generate leoric version from git tags on build
this introduces an intermediary build step which outputs artefacts in a
./build directory. The main reason here is so we could have a pre-install
step that inserts the version generated from git tags into the leoric
executable. The install step will now take the artefacts from ./build and
install it into the location from there.

This commit also includes a packaging fix. Previously `make packages` would
fail from a clean repository as there was no task to install files into ./usr
from which package contents are taken. This introduces this task and also
makes the packages task depend on it.