move detailed setup and macro explanation to man page
The more detailed explanations should be in the man page as it's likely that
those problems arise while using leoric. I removed it from the README so I
don't have to keep it in sync in 2 places and it makes the README less
overwhelming. The README now links to the online man page.