make VERSION injectable for source builds
when building from source downloads we don't have the git data available, so
we can't derive the version from tags there. In those scenarios, we allow the
buildsystem to inject the version as most likely you are downloading from a
tag anyways, so you know which version you are building.